AudioIntegration!!!!! Finally!!
Well, somewhat.  The Analyzer code now works concurrently with
AudioRunner because they are in seperate projects (they use
conflicting BadLogic library code).  We may need to integrate this to
the library later, but hopefully not.

We will follow this with streamlining the BLAudioAnalysis project and
making it do what Joseph actually wants it to.